data_generator_service/requirements.txt
pandas
faker
openai
python-dotenv
tqdm